+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Next version
 * Add Varanus::Domain
+* Add Varanus::SSL#list and Varanus::SSL#info
 
 ### 0.4.0 (2021-01-06)
 * Add Varanus::DCV